backend/requirements.txt
flask
flask-cors

python-dotenv
psutil

groq
gunicorn
requests

psycopg2
